// enums/mangaGenre.ts
export enum MangaGenre {
  ACTION = "Action",
  ADVENTURE = "Adventure",
  COMEDY = "Comedy",
  DEMONS = "Demons",
  DRAMA = "Drama",
  ECCHI = "Ecchi",
  FANTASY = "Fantasy",
  GENDER_BENDER = "Gender Bender",
  HAREM = "Harem",
  HISTORICAL = "Historical",
  HORROR = "Horror",
  JOSEI = "Josei",
  MAGIC = "Magic",
  MARTIAL_ARTS = "Martial Arts",
  MATURE = "Mature",
  MECHA = "Mecha",
  MILITARY = "Military",
  MYSTERY = "Mystery",
  ONE_SHOT = "One Shot",
  PSYCHOLOGICAL = "Psychological",
  ROMANCE = "Romance",
  SCHOOL_LIFE = "School Life",
  SCI_FI = "Sci Fi",
  SEINEN = "Seinen",
  SHOUJO = "Shoujo",
  SHOUJOAI = "Shoujoai",
  SHOUNEN = "Shounen",
  SHOUNENAI = "Shounenai",
  SLICE_OF_LIFE = "SliceOfLife",
  SMUT = "Smut",
  SPORTS = "Sports",
  SUPER_POWER = "SuperPower",
  SUPERNATURAL = "Supernatural",
  TRAGEDY = "Tragedy",
  VAMPIRE = "Vampire",
  YAOI = "Yaoi",
  YURI = "Yuri"
}